projects/preselector/pyproject.toml
[tool.poetry]
name = "preselector"
version = "0.1.0"
description = "A Python project by Cheffelo."
authors = ["Mats Eikeland Mollestad <mats.mollestad@cheffelo.com>"]
packages = [{include = "preselector"}]

[tool.poetry.dependencies]
python = "^3.10"
pydantic-settings = "^2.1.0"
data-contracts = {path = "../../packages/data-contracts", develop = true}
tqdm = "^4.66.2"
streamlit-pages = {path = "../../packages/streamlit-pages", develop = true}
pyodbc = {version = "4.0.35" }
aioodbc = {version = "0.4.1" }
sqlalchemy = "^2.0.30"
matplotlib = "^3.9.0"
fsspec = "^2024.6.1"
adlfs = "^2024.4.1"
sqlglot = "^25.7.1"
azure-keyvault-secrets = "^4.8.0"
polars-lts-cpu = "0.20.31"

[tool.poetry.group.worker.dependencies]
azure-servicebus = "^7.12.2"
cheffelo_logging = {path = "../../packages/cheffelo-logging", develop = true}

[tool.poetry.group.deploy.dependencies]
azure-mgmt-containerinstance = "^10.1.0"

[tool.poetry.group.databricks.dependencies]
databricks-env = {path = "../../packages/databricks-env", develop = true}

[tool.poetry.group.app.dependencies]
streamlit = "1.34.0"
pydantic-form = {path = "../../packages/pydantic-form", develop = true}

[tool.poetry.group.cli.dependencies]
pydantic-argparser = {path = "../../packages/pydantic-argparser", develop = true}

[tool.poetry.group.dev.dependencies]
pre-commit = "*"
pytest = "*"
pytest-asyncio = "*"
ruff = "*"
chef = {path = "../../packages/chef", develop = true}
pyright = "^1.1.376"
debugpy = "^1.8.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
exclude = [
    "**/jobs/**.py",
    "ui/deeplinks/show_results.py"
]

[tool.poetry.scripts]
chef = 'chef.cli:cli'
